requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
prometheus-client>=0.17.1
starlette-prometheus>=0.9.0